Rename pkg apis/oam to oam/core and update code generation
This renames the oam apis pkg to core to match the
group name of the CRDs. Also removes extraneous
generation methods that are not being used. This
resulted in a removal of a dependency on crossplane-
tools.
