+++ apis/core/v1alpha2/core_trait_types.go
