+++ cluster/charts/oam-kubernetes-runtime/crds/core.oam.dev_manualscalertraits.yaml
