+++ go.mod
@@ -4,7 +4,6 @@ go 1.13
 
 require (
 	github.com/crossplane/crossplane-runtime v0.7.1-0.20200421211018-be37c50cc2ab
-	github.com/crossplane/crossplane-tools v0.0.0-20200219001116-bb8b2ce46330
 	github.com/google/go-cmp v0.4.0
 	github.com/pkg/errors v0.8.1
 	github.com/prometheus/client_model v0.0.0-20190812154241-14fe0d1b01d4 // indirect
